Adding basic gotm-ersem systests
